+++ CHANGELOG.md
@@ -5,8 +5,7 @@ All notable changes to FrameLog will be documented in this file.
 ## [1.6.1-beta] - 2025-10-25
 
 ### Added
-- IsPrinting - checks whether it is possible to print with this minimum log level
-- SetMinimalLogLevel - sets the minimum log level
+- SetIsPrintingLoggerName - sets whether to print the logger name
 
 ### Fixed